# docs/swagger.yaml
definitions:
...
  faculty.UpdateFacultyRequestPayload:
    description: Information that should be available when you update a faculty
    properties:
      abbreviation:
        description: Faculty Name Abbreviation
        type: string
      id:
        description: Faculty ID, Provided by Query
        type: string
      name:
        description: Faculty Name
        type: string
      updateFacultyToken:
        description: Web Token that was appended to the link
        type: string
    required:
    - abbreviation
    - name
    type: object
...
  material.MaterialType:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - Handout
    - Video
    - External
  material.NewContentRequest:
    properties:
      link:
        type: string
      type:
        $ref: '#/definitions/material.MaterialType'
    required:
    - type
    type: object
  material.NewContentResponse:
    properties:
      upload_link:
        type: string
    type: object
...
  web.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        enum:
        - success
        - failed
        type: string
    type: object
info:
  contact: {}
  description: This is Open Couseware backend
  title: Open Courseware Application
  version: 1.0.1
paths:
...
  /course/faculty/{id}:
    get:
      description: Retrieve faculty data by UUID
      parameters:
      - description: Faculty ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "400":
          description: Invalid UUID provided in request path
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Get faculty by ID
      tags:
      - course
    patch:
      description: Update a faculty with the given ID
      parameters:
      - description: Faculty ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update Faculty Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/faculty.UpdateFacultyRequestPayload'
      - description: UpdateFacultyToken
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Update a faculty
      tags:
      - course
...
  /material/{id}:
    delete:
      consumes:
      - application/json
      description: Delete material
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Material id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Delete material
      tags:
      - content
    post:
      consumes:
      - application/json
      description: Add content of material
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add content request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/material.NewContentRequest'
      - description: Material id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/material.NewContentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Add Content
      tags:
      - content
...
swagger: "2.0"